Changed the data type of io_service_pool constructor argument from unsigned to size_t

+++ src/io_service_pool.cpp
@@ -1,10 +1,10 @@
 #include "io_service_pool.hpp"
 
-io_service_pool::io_service_pool(const unsigned concurrency) :
+io_service_pool::io_service_pool(const size_t concurrency) :
 	w(make_unique<work>(*this))
 {
 	reserve(concurrency);
-	for (int i = 0; i < concurrency; ++i)
+	for (size_t i = 0; i < concurrency; ++i)
 	{
 		emplace_back(async(launch::async, [&]()
 		{
